geoshop_engine/requirements.txt
requests
sqlalchemy
schedule
pymongo
python-dotenv
fastapi
uvicorn
pydantic
